Move the versioning functions to the new NWDoc class
